Add transition animation example. Place/tag lookup is more forgiving.
Places/tags aren't guaranteed to be in flow, because places are used for
two flows at once (currently) and also the final state of the animation
gets empty flow for uniformity (currently). May want to eventually
revert this change, so it's easier to tell when flow and node don't
agree on place/tag info.